backend/requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
bcrypt==4.3.0
Flask-Cors==5.0.1